Add integration test for packages
It adds a cloud build configuration file. When a tag is created, it
triggers a cloud build to run the integration test, which configures
gcloud, downloads kpt and anthoscli, installs the packages and cleans up
the cluster.
If the tests pass, a tag for anthoscli release will be generated based
on the current version text file and the file will be updated
accordingly.